repo_evaluator/requirements.txt
pandas
openpyxl
typer
openai
python-dotenv
pyyaml
